update: Replace old page structure view with tree
